Move helper function into `util::request` submodule
